+++ backend/api/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
-from .models import Team, Task
+from .models.task_models import Task
+from .models.team_models import Team
 
 admin.site.register(Team)
 admin.site.register(Task)
