+++ backend/api/models/team_models.py
@@ -1,0 +1,9 @@
+from django.db import models
+
+class Team(models.Model):
+    team_name = models.CharField(max_length=50)
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+    
+    def __str__(self):
+        return self.team_name
