+++ backend/api/serializers/serializers.py
@@ -1,7 +1,8 @@
 from django.contrib.auth.hashers import make_password
 from rest_framework import serializers
-from api.models import Task, Team
-from users.models.models import Users
+from api.models.team_models import Team
+from api.models.task_models import Task
+from users.models import Users
 
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
