+++ backend/users/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 from django.contrib.auth.models import AbstractUser
 from django.utils.translation import gettext_lazy as _
-from api.models.models import Team
+from api.models.team_models import Team
 
 class Users(AbstractUser):
     APPLICATION_STATUS = (
